# swagger/dist/build.yaml
openapi: 3.0.0
info:
  description: DXDATA의 스마트 관제시스템(Buoy) API입니다.
  version: 1.0.0
  title: Smart Control System API
  contact:
    email: dxdata@dxdata.co.kr
    url: 'http://www.dxdata.co.kr'
basePath: /
...
paths:
...
  /detail/group/list:
    get:
      tags:
        - detail
      description: 상세보기 화면에서 그룹 리스트를 보여줍니다.
      responses:
        '200':
          description: 각 그룹 리스트
          content:
            apllication/json:
              schema:
                $ref: '#/components/schemas/GroupList'
      security:
        - JWT: []
  /detail/group:
    get:
      tags:
        - detail
      description: |
        # 그룹의 라인별 상세 정보를 가져옵니다. 

        ## 라인별로 라인의 현재 정보와, 라인의 히스토리, 라인안에 있는 스마트 부표의 데이터들을 가져옵니다.

        ## 라인별로 배열로 데이터를 받습니다 _line_info 안에 라인 정보가 들어있습니다.

        ### 라인에 속해이는 부표들의 값들을 가져옵니다.

                "_buoy_list": [

                  {


                    "latitude": 좌표

                    "longitude": 좌표

                    "model": 모델 이름

                    "model_idx": 모델 번호

                    "salinity": 부표의 현재 염도 psu

                    "warn": 경고 여부

                    "water_temp": 부표의 현재 온도

                    "height": 높이

                    "weight": 부표의 현재 무게

                  },
               
                ],

        ### 라인 데이터의 7일치 기록을 가져옵니다.

                "_history": [
                  {
                    "date": 기록된 날짜입니다
                    "group_id": 속한 그룹의 id
                    "group_name": 속한 그룹의 이름
                    "latitude": 라인의 lat,
                    "longitude": 라인의 lng,
                    "line": 라인 번호,
                    "salinity": 평균 염분입니다 psu 단위입니다.,
                    "water_temp":평균 온도,
                    "weight": 43.529666900634766
                    "height": 평균 높이

                  }
                ],

        ### 현재 라인의 평균 정보 입니다.
              "_line_info" {
                "group_name": 라인이 속한 이름입니다.
                "line": 라인 번호입니다.,
                "latitude": 라인의 lat,
                "longitude": 라인의 lng,
                "salinity": 평균 염분입니다 psu 단위입니다.,
                "water_temp": 평균 온도,
                "weight": 평균 무게입니다.
                "height": 평균 높이입니다.
              }
            },
      parameters:
        - in: query
          name: group_id
          schema:
            type: integer
          example: 1
          required: true
      responses:
        '200':
          description: 그룹의 라인별 상세 정보
          content:
            apllication/json:
              schema:
                $ref: '#/components/schemas/GroupDetailLine'
      security:
        - JWT: []
  /detail/group/web:
    get:
      tags:
        - detail
      description: 'detail/group을 vuejs 받아오기 위해 데이터를 수정한 버전, 그룹의 평균 정보도 가져온다.'
      parameters:
        - in: query
          name: group_id
          schema:
            type: integer
          example: 1
          required: true
      responses:
        '200':
          description: 그룹의 라인별 상세 정보
          content:
            apllication/json:
              schema:
                $ref: '#/components/schemas/GroupDetailLine'
      security:
        - JWT: []
  /detail/group/history:
    get:
      tags:
        - detail
      description: group의 일주일간 기록을 가져옵니다.
      parameters:
        - in: query
          name: group_id
          schema:
            type: integer
          example: 1
          required: true
      responses:
        '200':
          description: Group의 일주일간 기록
          content:
            apllication/json:
              schema:
                $ref: '#/components/schemas/GroupHistory'
      security:
        - JWT: []
  /detail/group/modify:
    put:
      tags:
        - detail
      description: |
        group의 정보를 수정합니다. 
        group_system => 0 : 일반수하식, 1 : 연승 수하식, 2 : 뗏목, 3 : 기타
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupModify'
      responses:
        '200':
          description: '0이면 실패, 1이면 성공'
          content:
            apllication/json:
              schema:
                $ref: '#/components/schemas/Code'
      security:
        - JWT: []
  /detail/group/create:
    post:
      tags:
        - detail
      description: 새로운 그룹을 생성합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupAdd'
      responses:
        '200':
          description: '0이면 실패, 1이면 성공'
          content:
            apllication/json:
              schema:
                $ref: '#/components/schemas/Code'
      security:
        - JWT: []
  /detail/group/delete:
    post:
      tags:
        - detail
      description: '그룹을 삭제하고, 할당되어 있는 buoy들의 그룹을 해제합니다.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - group_id
              properties:
                group_id:
                  type: integer
              example:
                group_id: 8
      responses:
        '200':
          description: '0이면 실패, 1이면 성공'
          content:
            apllication/json:
              schema:
                $ref: '#/components/schemas/Code'
      security:
        - JWT: []
  /detail/buoy:
    get:
      tags:
        - detail
      description: |
        # 해당 부이의 상세정보를 가져옵니다.

              {
                  group_id: 그룹 id
                  group_name: 그룹 이름
                  line: 라인 번호
                  model: 부이 이름
                  model_idx: 부이 번호
                  latitude: 위도
                  longitude: 경도
                  salinity: 염분
                  water_temp: 수온
                  weight: 무게
                  height: 높이
                  warn: 경고 여부 0이면 없음 1이면 있음
                  warn_detail:
                    height_warn: 높이 경고 0: 없음, 1: 낮음, 2: 높음
                    location_warn: 위치 경고 0:없음, 1: 있음
                    salinity_warn: 염분 경고 0: 없음, 1: 낮음, 2: 높음
                    temp_warn: 수온 경고 0: 없음, 1: 낮음, 2: 높음
                    weight_warn: 무게 경고 0: 없음 1: 있음
              },
      parameters:
        - in: query
          name: model
          schema:
            type: string
          example: buoy_1
          required: true
      responses:
        '200':
          description: Buoy의 상세정보
          content:
            apllication/json:
              schema:
                $ref: '#/components/schemas/BuoyList'
      security:
        - JWT: []
  /detail/buoy/list:
    get:
      tags:
        - detail
      description: |
        # 그룹내의 모든 부이의 상세 정보를 가져옵니다.

              "_buoy_list": [

                {
                    group_id: 그룹 id
                    group_name: 그룹 이름
                    line: 라인 번호
                    model: 부이 이름
                    model_idx: 부이 번호
                    latitude: 위도
                    longitude: 경도
                    salinity: 염분
                    water_temp: 수온
                    weight: 무게
                    height: 높이
                    warn: 경고 여부 0이면 없음 1이면 있음
                    warn_detail:
                      height_warn: 높이 경고 0: 없음, 1: 낮음, 2: 높음
                      location_warn: 위치 경고 0:없음, 1: 있음
                      salinity_warn: 염분 경고 0: 없음, 1: 낮음, 2: 높음
                      temp_warn: 수온 경고 0: 없음, 1: 낮음, 2: 높음
                      weight_warn: 무게 경고 0: 없음 1: 있음

                },
            
              ],
      parameters:
        - in: query
          name: group_id
          schema:
            type: integer
          example: 1
          required: true
      responses:
        '200':
          description: Buoy 그룹별 리스트 및 정보
          content:
            apllication/json:
              schema:
                $ref: '#/components/schemas/BuoyList'
      security:
        - JWT: []
  /detail/buoy/history:
    get:
      tags:
        - detail
      description: Buoy의 일주일간 기록을 가져옵니다.
      parameters:
        - in: query
          name: model
          schema:
            type: string
          example: buoy_1
          required: true
      responses:
        '200':
          description: Buoy의 일주일간 기록
          content:
            apllication/json:
              schema:
                $ref: '#/components/schemas/BuoyHistory'
      security:
        - JWT: []
  /detail/buoy/assigned:
    get:
      tags:
        - detail
      description: 그룹 할당이 되어 있는 Buoy의 리스트를 가져옵니다.
      responses:
        '200':
          description: 할당되어 있는 부이 리스트
          content:
            apllication/json:
              schema:
                $ref: '#/components/schemas/AssignedBuoyList'
      security:
        - JWT: []
  /detail/buoy/unassigned:
    get:
      tags:
        - detail
      description: 유저에게는 할당되어있지만 그룹 할당이 되어 있지 않은 Buoy의 리스트를 가져옵니다.
      responses:
        '200':
          description: 할당되어 있지 않은 buoy 리스트
          content:
            apllication/json:
              schema:
                $ref: '#/components/schemas/UnAssignedBuoyList'
      security:
        - JWT: []
  /detail/buoy/allocate:
    put:
      tags:
        - detail
      description: Buoy의 그룹과 라인을 할당합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuoyAllocate'
      responses:
        '200':
          description: 성공 여부 코드
          content:
            apllication/json:
              schema:
                $ref: '#/components/schemas/Code'
      security:
        - JWT: []
  /detail/buoy/allocate/list:
    put:
      tags:
        - detail
      description: Buoy의 그룹과 라인을 할당합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/BuoyAllocate'
      responses:
        '200':
          description: 성공 여부 코드
          content:
            apllication/json:
              schema:
                $ref: '#/components/schemas/Code'
      security:
        - JWT: []
  /detail/buoy/deallocate:
    put:
      tags:
        - detail
      description: Buoy의 그룹과 라인 할당을 해제합니다..
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuoyDeallocate'
      responses:
        '200':
          description: 성공 여부 코드
          content:
            apllication/json:
              schema:
                $ref: '#/components/schemas/Code'
      security:
        - JWT: []
  /detail/buoy/deallocate/list:
    put:
      tags:
        - detail
      description: Buoy의 그룹과 라인 할당을 해제합니다..
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/BuoyDeallocate'
      responses:
        '200':
          description: 성공 여부 코드
          content:
            apllication/json:
              schema:
                $ref: '#/components/schemas/Code'
      security:
        - JWT: []
...
components:
  schemas:
...
    GroupList:
      type: object
      required:
        - group_id
        - group_name
      properties:
        group_id:
          type: integer
        group_name:
          type: string
      example:
        group_id: 1
        group_name: A
    GroupDetailLine:
      type: object
      required:
        - _buoy_list
        - _history
        - _line_info
      properties:
        _buoy_list:
          type: object
          properties:
            model:
              type: string
            model_idx:
              type: integer
            latitude:
              type: double
            longitude:
              type: double
            water_temp:
              type: float
            salinity:
              type: float
            height:
              type: float
            weight:
              type: float
            warn:
              type: integer
        _history:
          type: object
          properties:
            date:
              type: string
            group_id:
              type: integer
            group_name:
              type: string
            line:
              type: integer
            latitude:
              type: double
            longitude:
              type: double
            salinity:
              type: float
            water_temp:
              type: float
            height:
              type: float
            weight:
              type: float
        _line_info:
          type: object
          properties:
            group_name:
              type: string
            line:
              type: integer
            latitude:
              type: double
            longitude:
              type: double
            water_temp:
              type: float
            salinity:
              type: float
            height:
              type: float
            weight:
              type: float
      example:
        _buoy_list:
          model: buoy_26
          model_idx: 26
          latitude: 34.7971
          longitude: 128.4601
          salinity: 30.96930694580078
          water_temp: 12.993814468383789
          weight: 46.11586761474609
          height: 13.42252254486084
          warn: 0
        _history:
          date: '2022-04-20'
          group_id: 3
          group_name: C
          line: 1
          latitude: 32.3333
          longitude: 32.4444
          water_temp: 3.22
          salinity: 3.5
          height: 5.4
          weight: 2.3
        _line_info:
          group_name: A
          line: 1
          latitude: 34.7971
          longitude: 128.4603
          salinity: 31.11802520751953
          water_temp: 13.078953552246094
          weight: 43.428182983398436
          height: 13.810451889038086
    GroupHistory:
      type: object
      required:
        - date
        - group_id
        - group_name
        - group_latitude
        - group_longitude
        - group_water_temp
        - group_salinity
        - group_height
        - group_weight
      properties:
        date:
          type: string
        group_id:
          type: integer
        group_name:
          type: string
        group_latitude:
          type: double
        group_longitude:
          type: double
        group_water_temp:
          type: float
        group_salinity:
          type: float
        group_height:
          type: float
        group_weight:
          type: float
      example:
        date: '2022-04-27'
        group_id: 1
        group_name: A
        group_latitude: 32.3333
        group_longitude: 32.4444
        group_water_temp: 3.22
        group_salinity: 3.5
        group_height: 5.4
        group_weight: 2.3
    GroupModify:
      type: object
      required:
        - group_id
        - group_name
        - group_system
        - plain_buoy
      properties:
        group_id:
          type: integer
        group_name:
          type: string
        group_system:
          type: integer
        plain_buoy: integer
      example:
        group_id: 1
        group_name: A
        group_system: 0
        plain_buoy: 100
    GroupAdd:
      type: object
      required:
        - group_name
        - group_system
        - plain_buoy
      properties:
        group_name:
          type: string
        group_system:
          type: integer
        plain_buoy: integer
      example:
        group_name: ADD
        group_system: 0
        plain_buoy: 100
    BuoyList:
      type: object
      required:
        - group_id
        - group_name
        - model
        - model_idx
        - line
        - latitude
        - longitude
        - water_temp
        - salinity
        - height
        - weight
        - warn
      properties:
        group_id:
          type: integer
        group_name:
          type: string
        line:
          type: integer
        model:
          type: string
        model_idx:
          type: integer
        latitude:
          type: double
        longitude:
          type: double
        water_temp:
          type: float
        salinity:
          type: float
        height:
          type: float
        weight:
          type: float
        warn:
          type: integer
        warn_detail:
          type: object
          properties:
            height_warn:
              type: integer
            location_warn:
              type: integer
            salinity_warn:
              type: integer
            temp_warn:
              type: integer
            weight_warn:
              type: integer
      example:
        group_id: 1
        group_name: A
        line: 1
        model: buoy_1
        model_idx: 26
        latitude: 34.7971
        longitude: 128.4601
        salinity: 30.9693
        water_temp: 12.99381
        weight: 46.1158
        height: 13.4225
        warn: 0
        warn_detail:
          height_warn: 0
          location_warn: 0
          salinity_warn: 0
          temp_warn: 0
          weight_warn: 0
    BuoyHistory:
      type: object
      required:
        - avg_height
        - avg_salinity
        - avg_w_temp
        - avg_weight
        - date
        - model
      properties:
        avg_height:
          type: float
        avg_salinity:
          type: float
        avg_w_temp:
          type: float
        avg_weight:
          type: float
        date:
          type: string
        model:
          type: string
      example:
        avg_height: 14.394352
        avg_salinity: 31.537668
        avg_w_temp: 12.865586
        avg_weight: 42.267906
        date: 2022-04-19T00:00:00.000Z
        model: buoy_1
    AssignedBuoyList:
      type: object
      required:
        - model
        - model_idx
        - latitude
        - longitude
        - group_id
        - group_name
      properties:
        model:
          type: string
        model_idx:
          type: integer
        latitude:
          type: double
        longitude:
          type: double
        group_id:
          type: integer
        group_name:
          type: string
      example:
        model: buoy_1
        model_idx: 26
        latitude: 34.7971
        longitude: 128.4601
        group_id: 1
        group_name: A
    UnAssignedBuoyList:
      type: object
      required:
        - model
        - model_idx
        - line
        - latitude
        - longitude
      properties:
        model:
          type: string
        model_idx:
          type: integer
        latitude:
          type: double
        longitude:
          type: double
      example:
        model: buoy_1
        model_idx: 26
        latitude: 34.7971
        longitude: 128.4601
    BuoyAllocate:
      type: object
      required:
        - model
        - group_id
        - line
      properties:
        model:
          type: string
        group_id:
          type: integer
        line:
          type: integer
      example:
        model: buoy_101
        group_id: 4
        line: 1
    BuoyDeallocate:
      type: object
      required:
        - model
      properties:
        model:
          type: string
      example:
        model: buoy_101
...
    Code:
      type: object
      required:
        - code
      properties:
        code:
          type: integer
      example:
        code: 1
...
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT